+++ pages/index.js
@@ -6,7 +6,7 @@ import VideoReelLoop from '../components/video-reel-loop';
 
 export default function Home() {
   return (
-    <div className="text-imsetyBlack dark:text-imsetyWhite">
+    <div className="bg-imsetyWhite text-imsetyBlack dark:bg-imsetyBlack dark:text-imsetyWhite">
       <Head>
         <title>Imsety Taylor - Motion Designer</title>
         <meta
@@ -22,18 +22,18 @@ export default function Home() {
           rel="stylesheet"
         />
       </Head>
-      <main>
-        <div className="flex flex-col w-screen h-screen lg:flex-row">
-          <div className=" relative bg-imsetyWhite dark:bg-imsetyBlack p-8 md:p-20 flex flex-col justify-between h-1/2 lg:w-1/2 lg:h-full">
-            <Header />
+      <Header />
+      <main className="h-screen">
+        <div className="flex flex-col w-screen h-5/6  lg:flex-row">
+          <div className="container relative flex items-center h-1/2 lg:w-1/2 lg:h-full">
             <Hero />
-            <Footer />
           </div>
           <div className="relative h-1/2 lg:w-1/2 lg:h-full overflow-hidden">
             <VideoReelLoop />
           </div>
         </div>
       </main>
+      <Footer />
     </div>
   );
 }