+++ tailwind.config.js
@@ -1,10 +1,20 @@
 module.exports = {
   purge: ['./pages/**/*.{js,ts,jsx,tsx}', './components/**/*.{js,ts,jsx,tsx}'],
   darkMode: 'media', // or 'media' or 'class'
+  corePlugins: {
+    container: false
+  },
   theme: {
+    container: {
+      padding: {
+        DEFAULT: '2rem',
+        md: '5rem'
+      }
+    },
     fontFamily: {
       sans: ['Poppins', 'ui-sans-serif', 'system-ui']
     },
+
     extend: {
       colors: {
         imsetyBlack: '#131214',
@@ -13,7 +23,21 @@ module.exports = {
     }
   },
   variants: {
-    extend: {}
+    extend: {
+      backdropFilter: ['dark']
+    }
   },
-  plugins: []
+  plugins: [
+    ({ addComponents }) => {
+      addComponents({
+        '.container': {
+          paddingInline: '2rem',
+
+          '@screen md': {
+            paddingInline: '5rem'
+          }
+        }
+      });
+    }
+  ]
 };