+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- `Header` component with mobile navigation. (#12)
+- `Footer` component with a link to the [GitHub repo](https://github.com/imse-ty/imsety.com). (#12)
+- Default [container padding](https://tailwindcss.com/docs/container).
+
 ## [0.1.2]
 
 ### Added